requirements.txt
astroid==2.0.1
colorama==0.3.9
Django==2.0.7
django-ckeditor==5.6.1
django-js-asset==1.1.0
django-widget-tweaks==1.4.2
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pillow==5.2.0
pylint==2.0.0
pytz==2018.5
six==1.11.0
typed-ast==1.1.0
wrapt==1.10.11
whitenoise==4.1
gunicorn==19.9.0